+++ app/src/main/java/com/idormy/sms/forwarder/HelpActivity.java
@@ -7,10 +7,8 @@ import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 @SuppressWarnings("deprecation")
-public class HelpActivity extends AppCompatActivity {
+public class HelpActivity extends BaseActivity {
 
     @SuppressLint("SetJavaScriptEnabled")
     @Override
