+++ app/src/main/java/com/idormy/sms/forwarder/MainActivity.java
@@ -12,8 +12,6 @@ import android.os.Looper;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
@@ -23,7 +21,6 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.hjq.permissions.OnPermissionCallback;
@@ -54,12 +51,11 @@ import com.idormy.sms.forwarder.utils.TimeUtil;
 import com.idormy.sms.forwarder.view.StepBar;
 import com.umeng.commonsdk.UMConfigure;
 
-import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
 
 @SuppressWarnings("CommentedOutCode")
-public class MainActivity extends AppCompatActivity implements RefreshListView.IRefreshListener {
+public class MainActivity extends BaseActivity implements RefreshListView.IRefreshListener {
 
     private final String TAG = "MainActivity";
     // logVoList用于存储数据
@@ -487,61 +483,6 @@ public class MainActivity extends AppCompatActivity implements RefreshListView.I
         intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         intent.addCategory(Intent.CATEGORY_HOME);
         startActivity(intent);
-    }
-
-    //启用menu
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.menu_main, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    //menu点击事件
-    @SuppressLint("NonConstantResourceId")
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        Intent intent;
-        switch (item.getItemId()) {
-            case R.id.to_app_list:
-                intent = new Intent(this, AppListActivity.class);
-                break;
-            case R.id.to_clone:
-                intent = new Intent(this, CloneActivity.class);
-                break;
-            case R.id.to_about:
-                intent = new Intent(this, AboutActivity.class);
-                break;
-            case R.id.to_help:
-                //Uri uri = Uri.parse("https://gitee.com/pp/SmsForwarder/wikis/pages");
-                //intent = new Intent(Intent.ACTION_VIEW, uri);
-                intent = new Intent(this, HelpActivity.class);
-                break;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-
-        startActivity(intent);
-        return true;
-    }
-
-    //设置menu图标显示
-    @Override
-    public boolean onMenuOpened(int featureId, Menu menu) {
-        Log.d(TAG, "onMenuOpened, featureId=" + featureId);
-        if (menu != null) {
-            if (menu.getClass().getSimpleName().equals("MenuBuilder")) {
-                try {
-                    Method m = menu.getClass().getDeclaredMethod("setOptionalIconsVisible", Boolean.TYPE);
-                    m.setAccessible(true);
-                    m.invoke(menu, true);
-                } catch (NoSuchMethodException e) {
-                    Log.e(TAG, "onMenuOpened", e);
-                } catch (Exception e) {
-                    throw new RuntimeException(e);
-                }
-            }
-        }
-        return super.onMenuOpened(featureId, menu);
     }
 
     /*** 隐私协议授权弹窗*/
